requirements.txt
aiogram==2.23.1
python-dotenv==0.21.0
tortoise-orm==0.19.2